feat: Add PE cost function estimation

+++ src/ConfigurationManager.cpp
@@ -79,9 +79,9 @@ void loadConfiguration() {
     GlobalParams::traffic_distribution = readParam<string>(config, "traffic_distribution");
     GlobalParams::traffic_table_filename = readParam<string>(config, "traffic_table_filename");
     GlobalParams::clock_period_ps = readParam<int>(config, "clock_period_ps");
-    GlobalParams::simulation_time = readParam<int>(config, "simulation_time");
+    GlobalParams::simulation_time = readParam<unsigned long long>(config, "simulation_time");
     GlobalParams::n_virtual_channels = readParam<int>(config, "n_virtual_channels");
-    GlobalParams::reset_time = readParam<int>(config, "reset_time");
+    GlobalParams::reset_time = readParam<unsigned long long>(config, "reset_time");
     GlobalParams::stats_warm_up_time = readParam<int>(config, "stats_warm_up_time");
     GlobalParams::rnd_generator_seed = time(NULL);
     GlobalParams::detailed = readParam<bool>(config, "detailed");
